+++ packages/web/src/lib/view/parts-list/FilterPanel.spec.ts
@@ -7,11 +7,11 @@ import i18n from '$lib/i18n/define'
 import type { ACParts } from '@ac6_assemble_tool/parts/types/base/types'
 import type { CandidatesKey } from '@ac6_assemble_tool/parts/types/candidates'
 import { render, screen, fireEvent } from '@testing-library/svelte'
-import { describe, it, expect, beforeEach } from 'vitest'
+import { describe, it, expect } from 'vitest'
 
-import FilterPanel from '../FilterPanel.svelte'
-import { buildPropertyFilter } from './filters-application'
-import { numericOperands, type Filter } from './filters-core'
+import FilterPanel from './FilterPanel.svelte'
+import { buildPropertyFilter } from './state/filter/filters-application'
+import { numericOperands, type Filter } from './state/filter/filters-core'
 
 type FilterPanelProps = {
   slot: CandidatesKey
