+++ packages/web/src/lib/view/parts-list/state/state-serializer.ts
@@ -1,0 +1,106 @@
+import { type CandidatesKey } from '@ac6_assemble_tool/parts/types/candidates'
+import { logger } from '@ac6_assemble_tool/shared/logger'
+import { Result } from '@praha/byethrow'
+
+import { VALID_SLOTS, type DeserializeError } from './shared'
+import { type Filter } from './filter/filters-core'
+import { parseFilter } from './filter/serialization'
+import type { SortOrder } from './sort';
+import { parseSort } from './sort'
+
+// Filterは filters.ts からエクスポート
+export type { Filter } from './filter/filters-core'
+
+/**
+ * URL共有用の状態（スロット、フィルタ、並び替え）
+ */
+export interface SharedState {
+  slot: CandidatesKey
+  filters: Filter[]
+  sortKey: string | null
+  sortOrder: SortOrder | null
+}
+
+/**
+ * URLパラメータへのシリアライズ（共有用）
+ */
+export function serializeToURL(state: SharedState): URLSearchParams {
+  const params = new URLSearchParams()
+
+  // スロット
+  params.set('slot', state.slot)
+
+  // フィルタ条件（新しい型に対応）
+  for (const filter of state.filters) {
+    params.append('filter', filter.serialize())
+  }
+
+  // 並び替え
+  if (state.sortKey && state.sortOrder) {
+    params.set('sort', `${state.sortKey}:${state.sortOrder}`)
+  }
+
+  return params
+}
+/**
+ * URLパラメータからのデシリアライズ
+ */
+export function deserializeFromURL(
+  params: URLSearchParams,
+): Result.Result<SharedState, DeserializeError> {
+  try {
+    // スロット
+    const slotParam = params.get('slot')
+    let slot: CandidatesKey = 'rightArmUnit' // デフォルト
+
+    if (slotParam) {
+      if (!VALID_SLOTS.has(slotParam as CandidatesKey)) {
+        return Result.fail({ type: 'invalid_slot', slot: slotParam })
+      }
+      slot = slotParam as CandidatesKey
+    }
+
+    // フィルタ条件
+    const filterParams = params.getAll('filter')
+    const filters: Filter[] = []
+
+    for (const filterParam of filterParams) {
+      const parsed = parseFilter(filterParam)
+      if (parsed) {
+        filters.push(parsed)
+      } else {
+        logger.warn('Invalid filter parameter skipped', {
+          filter: filterParam,
+        })
+      }
+    }
+
+    // 並び替え
+    const sortParam = params.get('sort')
+    let sortKey: string | null = null
+    let sortOrder: SortOrder | null = null
+
+    if (sortParam) {
+      const parsed = parseSort(sortParam)
+      if (parsed) {
+        sortKey = parsed.key
+        sortOrder = parsed.order
+      } else {
+        logger.warn('Invalid sort parameter, using default', {
+          sort: sortParam,
+        })
+      }
+    }
+
+    return Result.succeed({ slot, filters, sortKey, sortOrder })
+  } catch (error) {
+    logger.error('Failed to deserialize state from URL', {
+      error: error instanceof Error ? error.message : String(error),
+    })
+
+    return Result.fail({
+      type: 'invalid_format',
+      message: error instanceof Error ? error.message : String(error),
+    })
+  }
+}
