+++ packages/web/src/lib/view/parts-list/state/view-mode.spec.ts
@@ -1,0 +1,46 @@
+import { describe, expect, it } from 'vitest'
+
+import { loadViewMode, saveViewMode } from './view-mode'
+
+describe('state/view-mode', () => {
+  describe('saveViewMode', () => {
+    it('表示モードをLocalStorageに保存できること', () => {
+      saveViewMode('list')
+
+      const saved = localStorage.getItem('ac6-parts-list-view-mode')
+      expect(saved).toBe('list')
+    })
+
+    it('表示モードを上書き保存できること', () => {
+      saveViewMode('grid')
+      saveViewMode('list')
+
+      const saved = localStorage.getItem('ac6-parts-list-view-mode')
+      expect(saved).toBe('list')
+    })
+  })
+
+  describe('loadViewMode', () => {
+    it('LocalStorageから表示モードを読み込めること', () => {
+      localStorage.setItem('ac6-parts-list-view-mode', 'list')
+
+      const viewMode = loadViewMode()
+
+      expect(viewMode).toBe('list')
+    })
+
+    it('LocalStorageに表示モードが存在しない場合、デフォルト(grid)を返すこと', () => {
+      const viewMode = loadViewMode()
+
+      expect(viewMode).toBe('grid')
+    })
+
+    it('LocalStorageの値が不正な場合、デフォルト(grid)を返すこと', () => {
+      localStorage.setItem('ac6-parts-list-view-mode', 'invalid')
+
+      const viewMode = loadViewMode()
+
+      expect(viewMode).toBe('grid')
+    })
+  })
+})
